Refactor code, add first part of events API
