+++ IIS/events/api_views.py
@@ -1,0 +1,10 @@
+from rest_framework import viewsets
+from rest_framework.decorators import action
+from rest_framework.response import Response
+
+from . import models, serializers
+
+
+class EventsViewSet(viewsets.ModelViewSet):
+    queryset = models.Event.objects.all()
+    serializer_class = serializers.EventSerializer
